Unifying db contexts
